requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0

# AWS integration
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
psycopg2-binary>=2.9,<3.0

# Redis (optional, for SSE pub/sub)
redis>=5.0,<6.0

# Celery (for background task execution)
celery>=5.3,<6.0
